double precision weights

+++ interface/HybridGBREvent.h
@@ -66,13 +66,13 @@
        unsigned char Class() const { return fClass; }
        float Var(int i) const { return fVars[i]; }
        unsigned short Quantile(int i) const { return fQuantiles[i]; }   
-       float Weight() const   { return fWeight;  }
+       double Weight() const   { return fWeight;  }
        
        void SetClass(unsigned char i) { fClass = i; }
        void SetVar(int i, float x) { fVars[i] = x; }
        void SetQuantile(int i, int q) { fQuantiles[i] = q; }
        //cache computed qunatities needed for split-search
-       void SetWeight(float x)     { fWeight = x; }
+       void SetWeight(double x)     { fWeight = x; }
        
        double Target(int i) const { return fTargets[i]; }
        float TransTarget(int i) const { return fTransTargets[i]; }
@@ -107,7 +107,7 @@
       double                   *fDerivatives2;
       unsigned int             *fCurrentNodes;
       double                    fPdfVal;
-      float                     fWeight;
+      double                    fWeight;
       unsigned char             fClass;
   };
